updated add to cart to increase quantity
