+++ src/pages/shop/shop.component.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import SHOP_DATA from './shop-data';
+import SHOP_DATA from './shop.data';
 import CollectionPreview from '../../components/collection-preview/collection-preview.component';
 
 class ShopPage extends React.Component {
@@ -16,13 +16,11 @@ class ShopPage extends React.Component {
 		const { collections } = this.state;
 		return (
 			<div className='shop-page'>
-				{
-					collections.map(({ id, ...otherCollectionProps }) =>
-						<CollectionPreview key={id} {...otherCollectionProps } />
-					)
-				}
+				{collections.map(({ id, ...otherCollectionProps }) => (
+					<CollectionPreview key={id} {...otherCollectionProps} />
+				))}
 			</div>
-		)
+		);
 	}
 }
