+++ src/redux/cart/cart.reducer.js
@@ -1,4 +1,5 @@
 import { CartActionTypes } from './cart.types';
+import { addItemToCart } from './cart.utils';
 
 const INITIAL_STATE = {
 	hidden: true,
@@ -15,7 +16,7 @@ const cartReducer = (state = INITIAL_STATE, action) => {
 		case CartActionTypes.ADD_ITEM:
 			return {
 				...state,
-				cartItems: [...state.cartItems, action.payload],
+				cartItems: addItemToCart(state.cartItems, action.payload),
 			};
 		default:
 			return state;
